seperate some funcs
